requirements.txt
# Testing Frameworks 
# Core Test Framework
pytest==8.3.3                     # 경량화된 테스트 프레임워크
pytest-xdist==3.6.1               # 병렬 실행 지원 (테스트 속도 향상)
pytest-rerunfailures==16.1        # 재시도 로직 

# API Testing
requests==2.32.3                   # API 테스트 및 HTTP 통신용
python-dotenv==1.0.1               # 환경변수(.env) 파일 관리

# E2E (UI) Testing
selenium==4.25.0                   # 웹 자동화 핵심 라이브러리
webdriver-manager==4.0.2           # 브라우저 드라이버 자동 관리

# Reporting / Quality (optional)
pytest-cov==6.0.0                  # 코드 커버리지 리포트
allure-pytest==2.15.0              #테스트 리포팅

# Logging (optional)
loguru==0.7.2                      #로깅(선택)
# colorama==0.4.6                  #콘솔 로그용 (선택)
